+++ packages/api/lib/applet.ts
@@ -1,13 +1,18 @@
 import { oc } from "@orpc/contract";
-// import { eventIterator } from "@orpc/server";
+import { eventIterator } from "@orpc/server";
 import { z } from "zod";
 
 
-export const AppletStateSchema = z.enum([
-    "running", 
-    "stopping", 
-    "dead",
+export const AppletState = z.enum([
+    "unknown",
+    "active",
+    "reloading",
+    "inactive",
+    "failed",
+    "activating",
+    "deactivating",
 ]);
+export type AppletState = z.infer<typeof AppletState>;
 
 export const AppletManagerAPIContract = {
     info: oc.output(
@@ -21,16 +26,12 @@ export const AppletManagerAPIContract = {
         })),
         destroy: oc.input(z.object({
             ref: z.string().optional(),
-        })),   
-        // subscribe: {
-        //     statechange: oc.output(
-        //         eventIterator(
-        //             z.object({
-        //                 ref: z.string(),
-        //                 state: AppletStateSchema,
-        //             })
-        //         )
-        //     ),
-        // },     
+        })),
+        getState: oc.input(z.object({
+            ref: z.string().optional(),
+        })).output(AppletState),
+        onStateChange: oc.input(z.object({
+            ref: z.string().optional(),
+        })).output(eventIterator(AppletState)),
     },
 };