+++ packages/server/lib/view.ts
@@ -13,6 +13,16 @@ export type ViewRequestHandler = Express.RequestHandler<
     {}, any, any, any, ViewLocals
 >;
 
+export interface ErrorViewLocals extends ViewLocals {
+    // TODO
+    error?: Error | any;
+}
+
+export type ErrorViewRequestHandler = Express.RequestHandler<
+    {}, any, any, any, ErrorViewLocals
+>;
+
 export interface IView extends ViewRequestHandler {
     onUnauthorized?: ViewRequestHandler;
+    onErrorDisplay?: ErrorViewRequestHandler;
 }